feat: add invoices to visits

+++ rest/WEB-INF/src/org/bandahealth/idempiere/rest/service/db/InvoiceLineDBService.java
@@ -1,14 +1,16 @@
 package org.bandahealth.idempiere.rest.service.db;
 
 import org.adempiere.exceptions.AdempiereException;
+import org.bandahealth.idempiere.base.model.MBHBPSpecificPayerInfo;
+import org.bandahealth.idempiere.base.model.MBHPayerInfoFld;
 import org.bandahealth.idempiere.base.model.MCharge_BH;
 import org.bandahealth.idempiere.base.model.MProduct_BH;
+import org.bandahealth.idempiere.rest.model.BusinessPartnerSpecificPayerInformation;
 import org.bandahealth.idempiere.rest.model.Charge;
 import org.bandahealth.idempiere.rest.model.InvoiceLine;
 import org.bandahealth.idempiere.rest.model.Product;
 import org.bandahealth.idempiere.rest.utils.DateUtil;
 import org.bandahealth.idempiere.rest.utils.StringUtil;
-import org.compiere.model.MElementValue;
 import org.compiere.model.MInvoiceLine;
 import org.compiere.model.Query;
 import org.compiere.util.Env;
@@ -16,6 +18,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -33,6 +37,12 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 	private AccountDBService accountDBService;
 	@Autowired
 	private ChargeDBService chargeDBService;
+	@Autowired
+	private OrderLineDBService orderLineDBService;
+	@Autowired
+	private BusinessPartnerSpecificPayerInformationDBService businessPartnerSpecificPayerInformationDBService;
+	@Autowired
+	private PayerInformationFieldDBService payerInformationFieldDBService;
 
 	@Override
 	public InvoiceLine saveEntity(InvoiceLine entity) {
@@ -46,6 +56,12 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 			invoiceLine.setC_Invoice_ID(entity.getInvoiceId());
 		}
 
+		if (entity.getOrderLineId() > 0) {
+			invoiceLine.setOrderLine(orderLineDBService.getEntityByIdFromDB(entity.getOrderLineId()));
+		} else if (entity.getOrderLine() != null && !StringUtil.isNullOrEmpty(entity.getOrderLine().getUuid())) {
+			invoiceLine.setOrderLine(orderLineDBService.getEntityByUuidFromDB(entity.getOrderLine().getUuid()));
+		}
+
 		if (entity.getCharge() != null && !StringUtil.isNullOrEmpty(entity.getCharge().getUuid())) {
 			MCharge_BH charge = chargeDBService.getEntityByUuidFromDB(entity.getCharge().getUuid());
 			if (charge != null) {
@@ -53,18 +69,17 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 			}
 		}
 
+		if (entity.getProduct() != null) {
+			MProduct_BH product = productDBService.getEntityByUuidFromDB(entity.getProduct().getUuid());
+			if (product != null) {
+				invoiceLine.setM_Product_ID(product.get_ID());
+			}
+		}
+
 		// All invoice lines need at least a charge or product, so error if nothing is
 		// there
 		if (invoiceLine.getC_Charge_ID() == 0 && invoiceLine.getM_Product_ID() == 0) {
 			throw new AdempiereException("Invoice Line missing a charge or product");
-		}
-
-		if (entity.getProduct() != null) {
-			MProduct_BH product = productDBService.getEntityByUuidFromDB(entity.getProduct().getUuid());
-
-			if (product != null) {
-				invoiceLine.setM_Product_ID(product.get_ID());
-			}
 		}
 
 		if (entity.getPrice() != null) {
@@ -87,6 +102,36 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 		invoiceLine.setDescription(entity.getDescription());
 
 		invoiceLine.saveEx();
+		entity.setId(invoiceLine.get_ID());
+
+		// If there is any information to save with this line, save it
+		if (entity.getBusinessPartnerSpecificPayerInformationList() != null) {
+			entity.setBusinessPartnerSpecificPayerInformationList(
+					entity.getBusinessPartnerSpecificPayerInformationList().stream().map(
+							businessPartnerSpecificPayerInformation -> {
+								businessPartnerSpecificPayerInformation.setInvoiceLineId(entity.getId());
+								return businessPartnerSpecificPayerInformationDBService.saveEntity(
+										businessPartnerSpecificPayerInformation);
+							}).collect(Collectors.toList()));
+		} else {
+			entity.setBusinessPartnerSpecificPayerInformationList(new ArrayList<>());
+		}
+		// Delete what is no longer there
+		List<MBHBPSpecificPayerInfo> businessPartnerSpecificPayerInformationList =
+				businessPartnerSpecificPayerInformationDBService.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+								MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, Collections.singleton(entity.getId()))
+						.get(entity.getId());
+		if (businessPartnerSpecificPayerInformationList != null) {
+			businessPartnerSpecificPayerInformationList.stream()
+					.filter(
+							existingBusinessPartnerSpecificPayerInformation -> entity.getBusinessPartnerSpecificPayerInformationList()
+									.stream()
+									.noneMatch(
+											newBusinessPartnerSpecificPayerInformation -> newBusinessPartnerSpecificPayerInformation.getUuid()
+													.equals(existingBusinessPartnerSpecificPayerInformation.getBH_BP_Specific_Payer_Info_UU())))
+					.forEach(orderLineChargeInformation -> businessPartnerSpecificPayerInformationDBService
+							.deleteEntity(orderLineChargeInformation.getBH_BP_Specific_Payer_Info_UU()));
+		}
 
 		return createInstanceWithAllFields(getEntityByUuidFromDB(invoiceLine.getC_InvoiceLine_UU()));
 	}
@@ -105,7 +150,7 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 						instance.getC_InvoiceLine_UU(), instance.isActive(), DateUtil.parse(instance.getCreated()),
 						instance.getCreatedBy(), instance.getC_Invoice_ID(),
 						new Product(product.getName(), product.getM_Product_UU(), product), instance.getPriceActual(),
-						instance.getQtyInvoiced(), instance.getLineNetAmt(), instance.getDescription());
+						instance.getQtyInvoiced(), instance.getLineNetAmt(), instance.getDescription(), instance);
 			} else if (instance.getC_Charge_ID() > 0) {
 				return new InvoiceLine(instance);
 			}
@@ -127,6 +172,17 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 				chargeDBService.transformData(new ArrayList<>(chargeDBService.getByIds(chargeIds).values())).stream()
 						.collect(Collectors.toMap(Charge::getId, charge -> charge));
 
+		// Batch call for insurance and donor payer information
+		Set<Integer> invoiceLineIds = dbModels.stream().map(MInvoiceLine::get_ID).collect(Collectors.toSet());
+		Map<Integer, List<MBHBPSpecificPayerInfo>> businessPartnerSpecificPayerInformationByOrderLineId =
+				businessPartnerSpecificPayerInformationDBService
+						.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+								MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, invoiceLineIds);
+		Map<Integer, MBHPayerInfoFld> payerInformationFieldsById = payerInformationFieldDBService.getByIds(
+				businessPartnerSpecificPayerInformationByOrderLineId.values().stream().flatMap(
+						businessPartnerSpecificPayerInformation -> businessPartnerSpecificPayerInformation.stream()
+								.map(MBHBPSpecificPayerInfo::getBH_Payer_Info_Fld_ID)).collect(Collectors.toSet()));
+
 		return dbModels.stream().map(invoiceLine -> {
 			InvoiceLine result = new InvoiceLine(invoiceLine);
 
@@ -136,6 +192,14 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 
 			if (chargesById.containsKey(invoiceLine.getC_Charge_ID())) {
 				result.setCharge(chargesById.get(invoiceLine.getC_Charge_ID()));
+			}
+			if (businessPartnerSpecificPayerInformationByOrderLineId.containsKey(result.getId())) {
+				result.setBusinessPartnerSpecificPayerInformationList(
+						businessPartnerSpecificPayerInformationByOrderLineId.get(result.getId()).stream()
+								.map(BusinessPartnerSpecificPayerInformation::new).peek(
+										businessPartnerSpecificPayerInformation -> businessPartnerSpecificPayerInformation.setPayerInformationFieldUuid(
+												payerInformationFieldsById.get(businessPartnerSpecificPayerInformation.getPayerInformationFieldId())
+														.getBH_Payer_Info_Fld_UU())).collect(Collectors.toList()));
 			}
 
 			return result;
@@ -168,6 +232,20 @@ public class InvoiceLineDBService extends BaseDBService<InvoiceLine, MInvoiceLin
 
 		List<MInvoiceLine> invoiceLines = new Query(Env.getCtx(), MInvoiceLine.Table_Name, whereClause, null)
 				.setParameters(invoiceId).setClient_ID().list();
+
+		// Get the associated order line charge information and delete it
+		Set<Integer> invoiceLineIds = invoiceLines.stream().map(MInvoiceLine::getC_InvoiceLine_ID)
+				.collect(Collectors.toSet());
+		boolean wereChildrenDeletesSuccessful = businessPartnerSpecificPayerInformationDBService
+				.getGroupsByIds(MBHBPSpecificPayerInfo::getC_InvoiceLine_ID,
+						MBHBPSpecificPayerInfo.COLUMNNAME_C_InvoiceLine_ID, invoiceLineIds)
+				.values().stream().flatMap(Collection::stream)
+				.allMatch(businessPartnerChargeInformation -> businessPartnerSpecificPayerInformationDBService
+						.deleteEntity(businessPartnerChargeInformation.getBH_BP_Specific_Payer_Info_UU()));
+		if (!wereChildrenDeletesSuccessful) {
+			throw new AdempiereException("There was an error deleting information");
+		}
+
 		for (MInvoiceLine invoiceLine : invoiceLines) {
 			invoiceLine.deleteEx(false);
 		}
